Split boss classes
