+++ EnemizerLibrary/Boss/BossType.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EnemizerLibrary
+{
+    public enum BossType
+    {
+        Kholdstare = 0,
+        Moldorm = 1,
+        Mothula = 2,
+        Vitreous = 3,
+        Helmasaur = 4,
+        Armos = 5,
+        Lanmola = 6,
+        Blind = 7,
+        Arrghus = 8,
+        Trixnexx = 9,
+        NoBoss = 255
+    }
+}
