+++ EnemizerLibrary/Boss/LanmolaBoss.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EnemizerLibrary
+{
+    public class LanmolaBoss : Boss
+    {
+        public LanmolaBoss() : base(BossType.Lanmola)
+        {
+            BossPointer = new byte[] { 0xCB, 0xDC };
+            BossGraphics = 11;
+        }
+    }
+}
